requirements.txt
flask
requests
twilio
python-dotenv
google-cloud-speech
google-cloud-texttospeech
google-cloud-storage
gunicorn
requests

